Allow per-Framework setting of defaults for table class and per page
This allows you to set defaults that will be used by all models.
It requires changing the Model API so that each model knows which
framework it is associated to, so if the model hasn't had a specific
setting, it can call into the framework to get the default.

Start adding unit tests, which just test model/framework API, and
remove some integration tests in favor of clearer unit tests.