Allow filters to have access to request information
This changes much of the API so that instead of passing type
strings to the model methods, Action instances are now passed,
which contain the type as well as a reference to the request.

The advantage to do this means that most of the model code can
now base actions based on properties of the request, allowing
for different behavior per-request instead of just per-type.

The disadvantage to doing this is that the API becomes more
difficult to unit test.