Handle many_to_one associations on forms
This changes the default columns used to automatically substitute
many_to_one associations for their respective foreign keys.

It makes it so that create/update/search uses the foreign key
values even when the column contains the association values,
and makes sure the model table uses the correct column values.