Have many to one lookup use eager/order/filter in associated model
Also, automatically use eager loading on search and browse pages
that will return the related models, including cascading the
eager loading for the associated model.

Currently this uses an ugly hack with a manually created action,
the surrounding API needs to be changed so that actions are not
passed, but requests are passed instead.