Reload Rails routes once per framework, not once per model
Reloading Rails routes is slow for large number of routes, so
reducing the number of times you do it should speed things up.

However, this means that if you add models outside the framework
block, routes won't be setup automatically.

In addition to only reloading once, only load a single route that
routes to all models.