Eagerly load associations of associated objects that use eager graph
In the search/browse tables, when automatically eager loading
associations because they are columns, if those associations
have a model class that eagerly graphs another association,
eagerly load that association as well.

This isn't a perfect fix, since some associations are specific
to graphing.  In such a case you need to make sure that for the
association type, eager uses [] so that eager_graph is not checked.

This also doesn't automatically handle the case where both eager
and eager_graph are used.