Add support for using flash string keys in the Roda support, to work with Roda's sessions plugin
This recognizes the :sessions_convert_symbols Roda application
option and if that is set then string keys are used.

This also adds automatically turns on support for using Roda's
sessions plugin in the tests if it is available. It also makes it
the default to run tests using the route_csrf plugin instead of
Rack::Csrf if Roda's sessions plugin is used.