Make autoforme.js file no longer require JQuery
This removes the implicit dependency on JQuery from autoforme.js.  As
the previous autocomplete library used depended on JQuery, this
switches to a new Javascript autocomplete library,
Pixabay/JavaScript-autoComplete.

The Javascript support now uses fetch, so a semi-modern browser is
now required.

This updates the demo site to use bootstrap 5.  The bootstrap 3
forme support doesn't currently work well with bootstrap 5, so drop
it from the demo site.