Drop CI support for Ruby 2.3 and JRuby 9.1
bundle install on JRuby 9.1 took 21 minutes, and still failed with
base64 issues in the Rails tests (previously, it failed in the
Sinatra tests).

Ruby 2.3 is still running bundle install (over 33 minutes now).

Tests pass locally, and passed in CI previously, so there is
some new gem version being picked up.  It's not worth the time
it would take to fully debug this, and early and later versions
work, so dropping support seems best.