Support pagination via filtering instead of offset using pagination_strategy :filter (requires unambiguous order)
Offset pagination is inefficient, especially for high page numbers,
since it needs to check and ignore everything before the offset.
Filter pagination can use database indexes, and can be fast even for
high page numbers.

Offset pagination does have the advantage in that it allows you to
click the Previous button, which is not supported with filter
pagination. However, considering users can use their back buttons,
this isn't a significant advantage.

Implementation-wise, this builds on the internals of Sequel's
Dataset#paged_each method, which also allows for a filter pagination
strategy.

While here, omit the display of Next/Previous inputs if the links are
not valid.  Previously, these were included inside a tag with a
disabled attribute, so clicking on the did not have an effect, but
they still looked like links (unless styled differently), which made
them appear broken.